fix(lib/util): use esm imports instead of dynamic imports
We were previously using dynamic `import`s from an array of plugins we
wanted to make it easier to add plugins in the future, while reducing
the verbosity that come with ESM imports.

While this was neat, dynamic imports are not treeshakeable and hence
would break the action during runtime as the bundle produced by rollup
still contained them dynamically.

We now use regular ESM imports in order to work around this.